feat: adding home initial logic - adding create post logic - adding get posts logic - adding remove post logic - adding ScreenState extensions

+++ app/src/main/java/dev/spooky/socialmediaapp/data/repository/AuthRepository.kt
@@ -49,7 +49,7 @@ class AuthRepositoryImpl(
             }
         if (response.status == HttpStatusCode.OK) {
             val body = response.body<UserInfoResponse>()
-            val info = body.run { UserInfo(id, displayName, email, null, body.createdAt.toLong()) }
+            val info = body.run { UserInfo(id, displayName, email, null, body.createdAt) }
             sessionHelper.setUserInfo(info)
             return Result.success(Unit)
         }
